Rename detail/bulk_invoke to detail/bulk_functions
