+++ agency/detail/bulk_functions/scope_result.hpp
