+++ testing/unorganized/test_shared.cpp
@@ -3,7 +3,7 @@
 #include <agency/detail/tuple.hpp>
 #include <agency/detail/index_cast.hpp>
 #include <agency/detail/shape_cast.hpp>
-#include <agency/detail/bulk_invoke/shared_parameter.hpp>
+#include <agency/detail/bulk_functions/shared_parameter.hpp>
 #include <agency/sequential_executor.hpp>
 #include <agency/execution_policy.hpp>
 #include <agency/detail/is_call_possible.hpp>
